refactor: enable by default
