+++ packages/preset-umi/src/index.ts
@@ -58,6 +58,7 @@ export default () => {
       require.resolve('./features/swc/swc'),
       require.resolve('./features/ui/ui'),
       require.resolve('./features/hmrGuardian/hmrGuardian'),
+      require.resolve('./features/firstRoutePreload/firstRoutePreload'),
 
       // commands
       require.resolve('./commands/build'),
