+++ packages/preset-umi/templates/firstRoutePreload/preloadRouteFilesScp.js
